Added a bunch of shit
Added our plugin project and built source so that the exe reference in
the project should be the same one I used. Plugin doesn't do anything,
compiles but I can't figure out how to read parts of entries, just how
to add them but that's useless to us rn. Also added the sample plugin
source but that doesn't have anything on reading entries either but
could be useful.